+++ lib/number/condition.hxx
@@ -6,10 +6,10 @@
 namespace exolix {
     class NumberCondition {
     public:
-        static bool isEven(__int128_t number);
-        static bool isOdd(__int128_t number);
+        static bool isEven(uint64_t number);
+        static bool isOdd(uint64_t number);
 
-        static bool isNumber(__int128_t number);
+        static bool isNumber(uint64_t number);
         static bool isNumber(const std::string &string);
     };
 }
