+++ lib/string/token.hxx
@@ -8,7 +8,7 @@ namespace exolix {
     public:
         static std::vector<std::string> split(const std::string &string, const std::string &delimiter);
 
-        static std::string toString(__int128_t number);
+        static std::string toString(int64_t number);
         static std::string toString(const std::string &string);
     };
 }
